Move test files to correct namespaces, add coverage tags
This moves all test classes to the namepsance and tests/* directory that
corresponds to the area of code they are testing, all supporting
classes to classes/tests, and adds @coverage tags to test classes.

This should make the test classes complaint with moodle-cs, and let core
tests autoload the test helper classes successfully.
