+++ classes/tests/test_azure_integration_client.php
