+++ classes/tests/test_file_system.php
@@ -28,13 +28,8 @@
 
 namespace tool_objectfs\tests;
 
-defined('MOODLE_INTERNAL') || die();
-
 use tool_objectfs\local\manager;
 use tool_objectfs\local\store\object_file_system;
-require_once(__DIR__ . '/test_client.php');
-require_once(__DIR__ . '/test_s3_integration_client.php');
-require_once(__DIR__ . '/test_azure_integration_client.php');
 
 class test_file_system extends object_file_system {
