+++ classes/tests/test_s3_integration_client.php
