+++ tests/local/object_manipulator/deleter_test.php
@@ -14,17 +14,16 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace tool_objectfs\tests;
-
-defined('MOODLE_INTERNAL') || die();
+namespace tool_objectfs\local\object_manipulator;
 
 use tool_objectfs\local\manager;
-use tool_objectfs\local\object_manipulator\deleter;
 
-require_once(__DIR__ . '/classes/test_client.php');
-require_once(__DIR__ . '/tool_objectfs_testcase.php');
-
-class deleter_test extends tool_objectfs_testcase {
+/**
+ * Tests for object deleter.
+ *
+ * @covers \tool_objectfs\local\object_manipulator\deleter
+ */
+class deleter_test extends \tool_objectfs\tests\testcase {
 
     /** @var string $manipulator */
     protected $manipulator = deleter::class;
