+++ tests/local/object_manipulator/puller_test.php
@@ -14,17 +14,16 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace tool_objectfs\tests;
-
-defined('MOODLE_INTERNAL') || die();
+namespace tool_objectfs\local\object_manipulator;
 
 use tool_objectfs\local\manager;
-use tool_objectfs\local\object_manipulator\puller;
 
-require_once(__DIR__ . '/classes/test_client.php');
-require_once(__DIR__ . '/tool_objectfs_testcase.php');
-
-class puller_test extends tool_objectfs_testcase {
+/**
+ * Tests for object puller.
+ *
+ * @covers \tool_objectfs\local\object_manipulator\puller
+ */
+class puller_test extends \tool_objectfs\tests\testcase {
 
     /** @var string $manipulator */
     protected $manipulator = puller::class;
